src/third_party/blink/web_tests/http/tests/devtools/portals/portals-elements-nesting-after-adoption-expected.txt

Tests that adopted portal is rendered inline correctly.


Running: testSetUp
- <html>
      <head></head>
    - <body>
        - <portal src="activate-predecessor.html">
            - #document
                - <html>
                      <head></head>
                    - <body>
                        - <script>
                              window.onportalactivate = e => {\n      var portal = e.adoptPredecessor();\n      portal.activate();\n    }\n
                          </script>
                      </body>
                  </html>
          </portal>
        - <script>
              async function waitForAdoption() {\n      var portal = document.querySelector("portal");\n      await portal.activate();\n      document.body.removeChild(portal);\n      return new Promise(resolve => {\n        window.onportalactivate = e => {\n          portal = e.adoptPredecessor();\n          document.body.appendChild(portal);\n          resolve();\n        }\n      });\n    }\n
          </script>
      </body>
  </html>

Running: testAfterAdoption
- <html>
      <head></head>
    - <body>
        - <script>
              async function waitForAdoption() {\n      var portal = document.querySelector("portal");\n      await portal.activate();\n      document.body.removeChild(portal);\n      return new Promise(resolve => {\n        window.onportalactivate = e => {\n          portal = e.adoptPredecessor();\n          document.body.appendChild(portal);\n          resolve();\n        }\n      });\n    }\n
          </script>
        - <portal>
            - #document
                - <html>
                      <head></head>
                    - <body>
                        - <script>
                              window.onportalactivate = e => {\n      var portal = e.adoptPredecessor();\n      portal.activate();\n    }\n
                          </script>
                      </body>
                  </html>
          </portal>
      </body>
  </html>
